requirements.txt
Flask
openai
Pillow
requests
python-dotenv
SQLAlchemy
